adds unit tests
