+++ src/router/index.js
@@ -1,28 +1,8 @@
 import Vue from 'vue';
 import VueRouter from 'vue-router';
-import ExpensesPage from '@/components/transaction/expenses-page.vue';
+import routes from './routes.js';
 
 Vue.use(VueRouter);
-
-const routes = [
-  {
-    path: '/',
-    name: 'Expenses',
-    component: ExpensesPage,
-  },
-  {
-    path: '/settings',
-    name: 'Settings',
-    // Generates separate chunk (settings.[hash].js), which is lazy-loaded when the route is visited.
-    component: () => import(/* webpackChunkName: "settings" */ '@/components/settings/settings-page.vue'),
-  },
-  {
-    path: '/about',
-    name: 'About',
-    // Generates separate chunk (about.[hash].js), which is lazy-loaded when the route is visited.
-    component: () => import(/* webpackChunkName: "about" */ '@/components/about/about-page.vue'),
-  },
-];
 
 const router = new VueRouter({
   mode: 'history',
